Code to produce gene plot
The R code produces the required gene plot requested by Dr Karl Berator.

The cleaned dataset was loaded, the required fonts and colours are then set up. The dataset was then separated into subsets for cell lines "Cell-type 101" and "Wild-type". The plots are then created, merged and saved as a .tiff file in the "pretty-figures" folder