Code written to calculate the required sample size
The R code calculates the required sample size for a linear regression to achieve a power of 90% at a significance level of 0.05 with a previously calculated R-squared of 0.1. There are 5 predictors in the model, including concentration, cell age, treatment (two levels), cell type (two levels), and media (two levels).

The effect size, f2, is calculated as R^2/(1-R^2)

From the output of the code, the number of error degrees, v is 147.8645, thus the sample size required is n = u + v + 1, where u is the number of predictors, in our case,
u=5. The required sample size is calculated as 153.8645, rounding off to 154.